backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
certifi==2025.8.3
charset-normalizer==3.4.3
fastapi==0.118.0
idna==3.10
neon-api==0.3.0
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
requests==2.32.5
sniffio==1.3.1
starlette==0.48.0
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.22.0
bcrypt==5.0.0
# python-jose[cryptography]==1.0.0
slowapi==0.1.9
passlib[argon2]==1.7.4
# transformers==4.57.0
PyJWT==2.8.0
orjson == 3.11.3